combining tests and pep8
