+++ RadClass/RadClass.py
@@ -90,7 +90,7 @@ class RadClass:
         '''
 
         # extract requisite data rows
-        data_matrix = np.empty((len(rows),len(self.cache[0])))
+        data_matrix = np.empty((len(rows), len(self.cache[0])))
         # normalize by live times to produce count rate data
         # processor.live can be indexed by appropriate timestamps but
         # data_matrix only has indices for rows.
@@ -103,8 +103,8 @@ class RadClass:
             data_matrix[i] = self.cache[idx] / dead_time
 
         # old, more inefficient way of summing
-        #total = np.zeros_like(data_matrix[0])
-        #for row in data_matrix:
+        # total = np.zeros_like(data_matrix[0])
+        # for row in data_matrix:
         #    total += row
         # utilizes numpy architecture to sum data
         total = np.sum(data_matrix, axis=0)
